UI family basic done!

+++ include/utils/FPSCounter.h
@@ -6,7 +6,7 @@ public:
     void StartFrame();
     void EndFrame();
     int GetFPS() const { return currentFPS_; }
-    int GetLastFrameTime() const { return lastFrameTime_; }
+    float GetLastFrameTime() const { return lastFrameTime_; }
 
 private:
     
@@ -14,5 +14,5 @@ private:
     std::chrono::steady_clock::time_point frameStartTime_;
     int frameCount_ = 0;
     int currentFPS_ = 0;
-    int lastFrameTime_ = 0;
+    float lastFrameTime_ = 0;
 };
